+++ src/fhirpath.d.ts
@@ -1,24 +1,16 @@
-export function compile(
+export function compile<T extends OptionVariants>(
   path: string | Path,
   model?: Model,
-  options?: {
-    resolveInternalTypes?: boolean
-    traceFn?: (value: any, label: string) => void,
-    userInvocationTable?: UserInvocationTable
-  }
-): Compile;
+  options?: T
+): Compile<T>;
 
-export function evaluate(
+export function evaluate<T extends OptionVariants>(
   fhirData: any,
   path: string | Path,
   context?: Context,
   model?: Model,
-  options?: {
-    resolveInternalTypes?: boolean,
-    traceFn?: (value: any, label: string) => void
-    userInvocationTable?: UserInvocationTable
-  }
-): any[];
+  options?: T
+): ReturnType<T>;
 
 export function resolveInternalTypes(value: any): any;
 
@@ -48,7 +40,33 @@ interface Model {
   };
 }
 
-type Compile = (resource: any, context?: Context) => any[];
+interface Options {
+    resolveInternalTypes?: boolean
+    traceFn?: (value: any, label: string) => void,
+    userInvocationTable?: UserInvocationTable,
+    terminologyUrl?: string
+}
+
+interface NoAsyncOptions extends Options {
+  async?: false;
+}
+
+interface AsyncOptions extends Options {
+  async: true;
+}
+
+interface AlwaysAsyncOptions extends Options {
+  async: 'always';
+}
+
+type OptionVariants = NoAsyncOptions | AsyncOptions | AlwaysAsyncOptions;
+
+type ReturnType<T> =
+    T extends AlwaysAsyncOptions ? Promise<any[]> :
+    T extends NoAsyncOptions ? any[] :
+    any[] | Promise<any[]>;
+
+type Compile<T> = (resource: any, context?: Context) => ReturnType<T>;
 
 type Context = void | Record<string, any>;